Bugfix in kernel/kernel.c
It probably didn't do much, since I didn't notice anything visual,
but when clearing the bottom line when scrolling it set all bytes to 0
rather than to empty spaces with a color. So the char to draw was \0,
the foreground color was black, and the background color was black.
So the wrong character was placed there, but it wasn't visible because
fg and bg was the same color.

This came from memset only taking one byte as a value and only setting
that one byte. In the future we may need to rework this, or make a
special memset function for cases where the value is larger than one
byte. This could for example be done by having the value be a void*
and having another parameter saying the size of the value then just
figure out the rest.